DELIA-41736 : ABR to consider buffer level along with throughput
Reason for change: Before ABR Ramp up/down , check the buffer level availability.
If enough buffer to play , not to do the ABR . Also configure the timeout value based
on buffer availability. This avoids curl timeout with poor network on higher profile files
Next change is to reset the abr cache buffer after profile change .
Test Procedure: Testcase added in ticket
Risks: Medium


DELIA-41736 : Provide configuration for enabling new ABR

Reason for change: Provide configuration setting to enable/disable new ABR
Buffer based ABR adds to existing throughput based functionality to keep steady state bitrate
Test Procedure: Testcase added in ticket
Risks: Medium


DELIA-41736 : Buffer based ABR estimation

Reason for change: Added additional logging to print buffer availability
Fixed the buffer calculation logic which got missed in last merge
Test Procedure: Testcase added in ticket
Risks: Low


DELIA-41736 : Disable buffer based ABR by default.

Reason for change: Force disabling buffer based ABR for this commit
This will override the App setting also . Added checks for position change for PTS change
Test Procedure: Testcase added in ticket
Risks: Low


DELIA-41736 : Added configurability for low and high buffer value

Reason for change: Added aamp.cfg option for abr low and high buffer value
These values are used for comparing during profile change
Test Procedure: Testcase added in ticket
Risks: Low


DELIA-41736 : Timeout change for segments and playlist.ABR fix

Reason for change: ABR rampdown on delayed download, which was not happening as
cache life was expired and ABR always returns same profile. Changed the timeout for
manifest and playlist . Retry is disabled if buffer duration is less than curltimeout

Test Procedure: Testcase added in ticket
Risks: Medium


DELIA-41736 : Prevent ABR rampdown when transition across discontinuity

Reason for change: Across discontinuity transistion ,only 2 fragments are cached ,
which causes the buffer lower than threshold point to trigger rampdown. This caused
Ad quality to go down.Prevent rampdown across discontinuity on low buffer

Test Procedure: Testcase added in ticket
Risks: Low

+++ test/aampcli.cpp
@@ -779,9 +779,9 @@ static void ProcessCliCommand(char *cmd)
 
 				case eAAMP_SET_NetworkTimeout:
                                 {
-                                        long networkTimeout;
+                                        double networkTimeout;
 					logprintf("Matchde Command eAAMP_SET_NetworkTimeout - %s ", cmd);
-					if (sscanf(cmd, "set %d %ld", &opt, &networkTimeout) == 2){
+					if (sscanf(cmd, "set %d %f", &opt, &networkTimeout) == 2){
 						mSingleton->SetNetworkTimeout(networkTimeout);
 					}
 					break;
